fix: make auto work from generate command
